requirements.txt
# Exam Research AI Agent - Dependencies
# Python 3.10+ recommended

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI & LLM
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.24
langchain-ollama>=0.2.0

# Vector Store
faiss-cpu>=1.7.4

# Web Scraping & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
ddgs>=9.0.0

# PDF handling (optional, for future PDF parsing)
pypdf==4.0.1

# Frontend
streamlit>=1.29.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging & Async
aiohttp==3.9.1
tenacity==8.2.3
supabase>=2.3.0
python-dotenv>=1.0.0
